requirements.txt
streamlit
librosa
matplotlib
numpy
pandas
Pillow
torch
torchvision
faiss-cpu
google-generativeai